fix: Entity props show evaluated value - Header bg color white
